Got Sphere working decently.
git-svn-id: https://10.0.0.20/svn/projects/Firefly_Algorithm@1521 cea9dca5-4048-41bc-958f-9b301c446f7b

+++ src/fa/Firefly.java
@@ -6,39 +6,47 @@ import java.util.Random;
  * @author Michael Lohrer
  * @version 0.1
  */
-public class Firefly implements FAConstants {
-	
-	private double lightIntensity;
-	private double[] position;
-	
+public class Firefly implements FAConstants
+{
+
+	private float lightIntensity;
+	private float[] position;
+
 	/**
 	 * @param numDimensions
 	 */
-	public Firefly(int numDimensions) {
-		position = new double[numDimensions];
+	public Firefly(int numDimensions)
+	{
+		position = new float[numDimensions];
 		Random r = new Random();
-		for (int i = 0; i < numDimensions; i++) {
-			position[i] = (r.nextDouble()-0.5)*BOUND;
+		for (int i = 0; i < numDimensions; i++)
+		{
+			position[i] = (r.nextFloat() - 0.5f) * BOUND;
 		}
 	}
-	
-	public double[] getPosition() {
+
+	public float[] getPosition()
+	{
 		return position;
 	}
-	
-	public void setPosition(double[] newPosition) {
+
+	public void setPosition(float[] newPosition)
+	{
 		position = newPosition;
 	}
-	
-	public void setPosition(double newPosition, int dimension) {
+
+	public void setPosition(float newPosition, int dimension)
+	{
 		position[dimension] = newPosition;
 	}
-	
-	public double getIntensity() {
+
+	public float getIntensity()
+	{
 		return lightIntensity;
 	}
-	
-	public void setIntensity(double intensity) {
+
+	public void setIntensity(float intensity)
+	{
 		lightIntensity = intensity;
 	}
 }
